requirements.txt
python-pptx==0.6.21
openai==1.12.0
python-dotenv==1.0.0
flask==2.3.3
requests==2.31.0 